+++ CHANGELOG.md
@@ -1483,3 +1483,4 @@
 - [2026-05-26 22:34] feat(network): add bi-directional event bus handler for network
 - [2026-05-26 22:37] fix(perf): fix memory leak by cancelling perf subscription
 - [2026-05-27 09:40] test(perf): implement widget integration tests for perf
+- [2026-05-27 09:52] perf(cache): optimize memory consumption in cache list viewport